ChangedToEverySecondPayouts
Made it so the game pays out every second, rather than every 30 seconds. Also, modified the game to allow the player to set game speed to 0 by pressing P, so they can calmly build the buildings they want when they have the funds to. This is to test a different style of game play to see if it feels better.